Updated to Include GAM and Some Model Comparison
Still sort of a rough draft. The method sort of changed half way through so some sections may be better suited for Samhit's part. He can just copy/paste + add best tree method for the table showing all the comparison metrics.

Additionally, some parts of the Methodology section will need to be adjusted given the change in strategy. 

Some of this stuff in the beginning will have to be shifted to the EDA section.